requirements.txt
pandas>=0.24.0
numpy>=1.22.2
requests>=2.26
multitasking>=0.0.7
lxml>=4.5.1
yahoofinancials>=1.6
pandas_datareader>=0.4.0
mysql-connector>=2.2.9
flask>=2.0.2
npm>=0.1.1
mysqlclient>=2.1.0
requests>=1.26.8
Flask-Cors>=3.0.10
pprintpp>=0.4.0
flask-mysqldb>=0.2.0
xlrd>=2.0.1
openpyxl>=3.0.9
matplotlib>=3.5.1
sqlalchemy>=1.4.31
yfinance>=0.1.70
pymysql>=1.0.2
